% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitDat.R
\name{split_dataset}
\alias{split_dataset}
\title{Split a dataset into training and holdout}
\usage{
split_dataset(data, ntrain, seed)
}
\arguments{
\item{'data'}{is a data frame or tbl for the dataset
'ntrain' will be numeric, 'seed' will be the numeric}
}
\value{
a list containing both training and holdout sets
}
\description{
Split a dataset into training and holdout
}
\details{
split a dataset into training and holdout sets
where the training set will contain 'ntrain' of examples
'ntrain' must be smaller or equal to the total size of 'data'
}
\examples{
split_dataset(datateachr::cancer_sample, 300, 447)
}
